Set correct policy-container for worker construction
This makes sure that when workers are created, their
global scope has the correct policy-container set
so that we can do CSP-checks.
