--- tests/wpt/meta/content-security-policy/connect-src/worker-connect-src-blocked.sub.html.ini
@@ -1,3 +1,0 @@
-[worker-connect-src-blocked.sub.html]
-  [Expecting logs: ["xhr blocked","TEST COMPLETE"\]]
-    expected: FAIL
