+++ tests/wpt/meta/content-security-policy/inside-worker/dedicatedworker-connect-src.html.ini
@@ -1,14 +1,5 @@
 [dedicatedworker-connect-src.html]
-  [Cross-origin 'fetch()' in http: with connect-src 'self']
-    expected: FAIL
-
-  [Cross-origin XHR in http: with connect-src 'self']
-    expected: FAIL
-
   [Same-origin => cross-origin 'fetch()' in http: with connect-src 'self']
-    expected: FAIL
-
-  [WebSocket in http: with connect-src 'self']
     expected: FAIL
 
   [Reports match in http: with connect-src 'self']
