+++ tests/wpt/meta/content-security-policy/inside-worker/dedicatedworker-script-src.html.ini
@@ -1,15 +1,16 @@
 [dedicatedworker-script-src.html]
+  expected: TIMEOUT
   [Cross-origin `importScripts()` blocked in http: with script-src 'self']
-    expected: FAIL
+    expected: TIMEOUT
 
   [`eval()` blocked in http: with script-src 'self']
-    expected: FAIL
+    expected: NOTRUN
 
   [`setTimeout([string\])` blocked in http: with script-src 'self']
-    expected: FAIL
+    expected: NOTRUN
 
   [Reports are sent for http: with script-src 'self']
-    expected: FAIL
+    expected: NOTRUN
 
   [Cross-origin `importScripts()` blocked in blob: with script-src 'self']
     expected: FAIL
@@ -22,3 +23,6 @@
 
   [Reports are sent for blob: with script-src 'self']
     expected: FAIL
+
+  [dedicatedworker-script-src]
+    expected: TIMEOUT
