--- tests/wpt/meta/content-security-policy/script-src/worker-eval-blocked.sub.html.ini
@@ -1,3 +1,0 @@
-[worker-eval-blocked.sub.html]
-  [Expecting logs: ["eval blocked"\]]
-    expected: FAIL
