--- tests/wpt/meta/content-security-policy/script-src/worker-function-function-blocked.sub.html.ini
@@ -1,3 +1,0 @@
-[worker-function-function-blocked.sub.html]
-  [Expecting logs: ["Function() function blocked"\]]
-    expected: FAIL
