+++ tests/wpt/meta/content-security-policy/script-src/worker-importscripts.sub.html.ini
@@ -1,6 +1,3 @@
 [worker-importscripts.sub.html]
-  [Dedicated worker delivers its own CSP]
-    expected: FAIL
-
   [Shared worker delivers its own CSP]
     expected: FAIL
