+++ tests/wpt/meta/content-security-policy/securitypolicyviolation/inside-dedicated-worker.html.ini
@@ -1,7 +1,4 @@
 [inside-dedicated-worker.html]
   expected: TIMEOUT
-  [SecurityPolicyViolation event fired on global.]
-    expected: FAIL
-
   [SecurityPolicyViolation event fired on global with the correct blockedURI.]
     expected: TIMEOUT
