+++ tests/wpt/meta/content-security-policy/wasm-unsafe-eval/default-src-blocks-wasm.any.js.ini
@@ -1,7 +1,4 @@
 [default-src-blocks-wasm.any.worker.html]
-  [default-src-blocks-wasm]
-    expected: FAIL
-
 
 [default-src-blocks-wasm.any.html]
