+++ tests/wpt/meta/content-security-policy/wasm-unsafe-eval/script-src-blocks-wasm.any.js.ini
@@ -2,9 +2,6 @@
   expected: ERROR
 
 [script-src-blocks-wasm.any.worker.html]
-  [script-src-blocks-wasm]
-    expected: FAIL
-
 
 [script-src-blocks-wasm.any.html]
