+++ tests/wpt/meta/content-security-policy/wasm-unsafe-eval/script-src-spv-asynch.any.js.ini
@@ -7,9 +7,8 @@
 
 
 [script-src-spv-asynch.any.worker.html]
-  expected: TIMEOUT
   [Securitypolicyviolation event looks like it should]
-    expected: TIMEOUT
+    expected: FAIL
 
 
 [script-src-spv-asynch.any.serviceworker.html]
