--- tests/wpt/meta/fetch/api/policies/csp-blocked-worker.html.ini
@@ -1,3 +1,0 @@
-[csp-blocked-worker.html]
-  [Fetch is blocked by CSP, got a TypeError]
-    expected: FAIL
