+++ tests/wpt/meta/trusted-types/should-sink-type-mismatch-violation-be-blocked-by-csp-002-worker.html.ini
@@ -1,4 +1,6 @@
 [should-sink-type-mismatch-violation-be-blocked-by-csp-002-worker.html]
-  expected: TIMEOUT
   [Checking reported violations for setTimeout(';;;;;') from DedicatedWorker]
-    expected: TIMEOUT
+    expected: FAIL
+
+  [Location of required-trusted-types-for violations.]
+    expected: FAIL
