+++ tests/wpt/meta/trusted-types/should-trusted-type-policy-creation-be-blocked-by-csp-004-worker.html.ini
@@ -1,10 +1,6 @@
 [should-trusted-type-policy-creation-be-blocked-by-csp-004-worker.html]
-  expected: TIMEOUT
-  [No violation/exception for allowed policy names (tt-policy-name-1 tt-policy-name-2 tt-policy-name-3).]
-    expected: TIMEOUT
+  [Exception and violations for CSP with multiple enforce and report-only policies.]
+    expected: FAIL
 
-  [Violation and exception for duplicate policy names (tt-policy-name-1 tt-policy-name-2 tt-policy-name-3).]
-    expected: NOTRUN
-
-  [Violation and exception for forbidden policy name (tt-policy-name-1 tt-policy-name-2 tt-policy-name-3).]
-    expected: NOTRUN
+  [Location of trusted-types violations.]
+    expected: FAIL
