+++ tests/wpt/meta/trusted-types/trusted-types-reporting-check-report-DedicatedWorker-sink-mismatch.html.ini
@@ -1,7 +1,4 @@
 [trusted-types-reporting-check-report-DedicatedWorker-sink-mismatch.html]
-  [Passing a plain string to eval throws.]
-    expected: FAIL
-
   [Test report-uri works with require-trusted-types-for violation.]
     expected: FAIL
