+++ tests/wpt/meta/trusted-types/trusted-types-reporting-for-DedicatedWorker-DedicatedWorker-constructor.html.ini
@@ -1,7 +1,3 @@
 [trusted-types-reporting-for-DedicatedWorker-DedicatedWorker-constructor.html]
-  expected: TIMEOUT
-  [No violation reported for Worker constructor with TrustedScriptURL.]
-    expected: TIMEOUT
-
   [Violation report for Worker constructor with plain string.]
-    expected: NOTRUN
+    expected: FAIL
