+++ tests/wpt/meta/trusted-types/trusted-types-reporting-for-DedicatedWorker-ServiceWorkerContainer-register.https.html.ini
@@ -1,7 +1,6 @@
 [trusted-types-reporting-for-DedicatedWorker-ServiceWorkerContainer-register.https.html]
-  expected: TIMEOUT
   [No violation reported for ServiceWorkerContainer register with TrustedScriptURL.]
-    expected: TIMEOUT
+    expected: FAIL
 
   [Violation report for ServiceWorkerContainer register with plain string.]
-    expected: NOTRUN
+    expected: FAIL
