+++ tests/wpt/meta/trusted-types/trusted-types-reporting-for-DedicatedWorker-eval.html.ini
@@ -1,7 +1,6 @@
 [trusted-types-reporting-for-DedicatedWorker-eval.html]
-  expected: TIMEOUT
   [No violation reported for eval with TrustedScript.]
-    expected: TIMEOUT
+    expected: FAIL
 
   [Violation report for eval with plain string.]
-    expected: NOTRUN
+    expected: FAIL
