+++ tests/wpt/meta/trusted-types/trusted-types-reporting-for-DedicatedWorker-function-constructor.html.ini
@@ -1,37 +1,12 @@
 [trusted-types-reporting-for-DedicatedWorker-function-constructor.html]
-  expected: TIMEOUT
-  [No violation reported for Function with TrustedScript.]
-    expected: TIMEOUT
-
-  [No violation reported for Function with multiple TrustedScript args.]
-    expected: NOTRUN
-
   [Violation report for Function with plain string.]
-    expected: NOTRUN
-
-  [No violation reported for AsyncFunction with TrustedScript.]
-    expected: NOTRUN
-
-  [No violation reported for AsyncFunction with multiple TrustedScript args.]
-    expected: NOTRUN
+    expected: FAIL
 
   [Violation report for AsyncFunction with plain string.]
-    expected: NOTRUN
-
-  [No violation reported for GeneratorFunction with TrustedScript.]
-    expected: NOTRUN
-
-  [No violation reported for GeneratorFunction with multiple TrustedScript args.]
-    expected: NOTRUN
+    expected: FAIL
 
   [Violation report for GeneratorFunction with plain string.]
-    expected: NOTRUN
-
-  [No violation reported for AsyncGeneratorFunction with TrustedScript.]
-    expected: NOTRUN
-
-  [No violation reported for AsyncGeneratorFunction with multiple TrustedScript args.]
-    expected: NOTRUN
+    expected: FAIL
 
   [Violation report for AsyncGeneratorFunction with plain string.]
-    expected: NOTRUN
+    expected: FAIL
