+++ tests/wpt/meta/trusted-types/trusted-types-reporting-for-DedicatedWorker-importScripts.html.ini
@@ -1,7 +1,3 @@
 [trusted-types-reporting-for-DedicatedWorker-importScripts.html]
-  expected: TIMEOUT
-  [No violation reported for importScripts with TrustedScriptURL.]
-    expected: TIMEOUT
-
   [Violation report for importScripts with plain string.]
-    expected: NOTRUN
+    expected: FAIL
