+++ tests/wpt/meta/trusted-types/trusted-types-reporting-for-DedicatedWorker-setTimeout-setInterval.html.ini
@@ -1,13 +1,6 @@
 [trusted-types-reporting-for-DedicatedWorker-setTimeout-setInterval.html]
-  expected: TIMEOUT
-  [No violation reported for setTimeout with TrustedScript.]
-    expected: TIMEOUT
-
-  [No violation reported for setInterval with TrustedScript.]
-    expected: NOTRUN
-
   [Violation report for setTimeout with plain string.]
-    expected: NOTRUN
+    expected: FAIL
 
   [Violation report for setInterval with plain string.]
-    expected: NOTRUN
+    expected: FAIL
